+++ src/backends/mos6502/instruction_set/tests/instructions.rs
@@ -1,11 +1,12 @@
-use crate::instruction_set::address_mode::AddressMode;
-use crate::instruction_set::mnemonics::Mnemonic;
-use crate::instruction_set::StaticInstruction;
+use crate::backends::mos6502::instruction_set::address_mode::AddressMode;
+use crate::backends::mos6502::instruction_set::mnemonics::Mnemonic;
+use crate::backends::mos6502::instruction_set::StaticInstruction;
+use crate::Emitter;
 
 #[test]
 fn instruction_with_accumulator_address_mode_should_return_single_byte() {
     let inst = StaticInstruction::new(Mnemonic::ASL, AddressMode::Accumulator);
-    let op: Vec<u8> = inst.into();
+    let op: Vec<u8> = inst.emit();
 
     assert_eq!(op, vec![0x0a])
 }
@@ -13,7 +14,7 @@ fn instruction_with_accumulator_address_mode_should_return_single_byte() {
 #[test]
 fn instruction_with_single_byte_operand_should_order_instructions_correctly() {
     let inst = StaticInstruction::new(Mnemonic::CPY, AddressMode::Immediate(0x12));
-    let op: Vec<u8> = inst.into();
+    let op: Vec<u8> = inst.emit();
 
     assert_eq!(op, vec![0xc0, 0x12])
 }
@@ -21,7 +22,7 @@ fn instruction_with_single_byte_operand_should_order_instructions_correctly() {
 #[test]
 fn instruction_with_two_byte_operand_should_order_operands_after_opcode_in_little_endian_format() {
     let inst = StaticInstruction::new(Mnemonic::CPY, AddressMode::Absolute(0x1234));
-    let op: Vec<u8> = inst.into();
+    let op: Vec<u8> = inst.emit();
 
     assert_eq!(op, vec![0xcc, 0x34, 0x12])
 }
