+++ src/backends/mos6502/parser/tests/address_mode.rs
@@ -1,5 +1,5 @@
-use crate::instruction_set::{AddressMode, Mnemonic};
-use crate::parser::instruction;
+use crate::backends::mos6502::instruction_set::{AddressMode, Mnemonic};
+use crate::backends::mos6502::parser::instruction;
 use parcel::prelude::v1::*;
 use parcel::MatchStatus;
 
@@ -8,9 +8,11 @@ macro_rules! gen_am_test {
         assert_eq!(
             Ok(MatchStatus::Match((
                 &$input[$input.len()..],
-                $crate::instruction_set::InstructionOrDefinition::Instruction(
-                    $crate::instruction_set::Instruction::from(
-                        $crate::instruction_set::StaticInstruction::new($mnemonic, $am)
+                $crate::backends::mos6502::instruction_set::InstructionOrDefinition::Instruction(
+                    $crate::backends::mos6502::instruction_set::Instruction::from(
+                        $crate::backends::mos6502::instruction_set::StaticInstruction::new(
+                            $mnemonic, $am
+                        )
                     )
                 )
             ))),
